+++ ansible_collections/community/postgresql/plugins/modules/postgresql_set.py
@@ -25,7 +25,7 @@ description:
 options:
   name:
     description:
-    - Name of PostgreSQL server parameter.
+    - Name of PostgreSQL server parameter. Pay attention that parameters are case sensitive (see examples below).
     type: str
     required: true
   value:
@@ -62,6 +62,7 @@ options:
     version_added: '0.2.0'
 notes:
 - Supported version of PostgreSQL is 9.4 and later.
+- Supports C(check_mode).
 - Pay attention, change setting with 'postmaster' context can return changed is true
   when actually nothing changes because the same value may be presented in
   several different form, for example, 1024MB, 1GB, etc. However in pg_settings
@@ -102,7 +103,8 @@ EXAMPLES = r'''
     value: 32mb
   register: set
 
-- ansible.builtin.debug:
+- name: Print the result if the setting changed
+  ansible.builtin.debug:
     msg: "{{ set.name }} {{ set.prev_val_pretty }} >> {{ set.value_pretty }} restart_req: {{ set.restart_required }}"
   when: set.changed
 # Ensure that the restart of PostgreSQL server must be required for some parameters.
@@ -118,6 +120,12 @@ EXAMPLES = r'''
   community.postgresql.postgresql_set:
     name: wal_log_hints
     value: default
+
+- name: Set TimeZone parameter (careful, case sensitive)
+  community.postgresql.postgresql_set:
+    name: TimeZone
+    value: 'Europe/Paris'
+
 '''
 
 RETURN = r'''
@@ -180,7 +188,7 @@ from ansible.module_utils._text import to_native
 PG_REQ_VER = 90400
 
 # To allow to set value like 1mb instead of 1MB, etc:
-POSSIBLE_SIZE_UNITS = ("mb", "gb", "tb")
+LOWERCASE_SIZE_UNITS = ("mb", "gb", "tb")
 
 # ===========================================
 # PostgreSQL module specific support methods.
@@ -198,6 +206,11 @@ def param_get(cursor, module, name):
 
     except Exception as e:
         module.fail_json(msg="Unable to get %s value due to : %s" % (name, to_native(e)))
+
+    if not info:
+        module.fail_json(msg="No such parameter: %s. "
+                             "Please check its spelling or presence in your PostgreSQL version "
+                             "(https://www.postgresql.org/docs/current/runtime-config.html)" % name)
 
     raw_val = info[0][1]
     unit = info[0][2]
@@ -233,32 +246,61 @@ def pretty_to_bytes(pretty_val):
     # if the value contains 'B', 'kB', 'MB', 'GB', 'TB'.
     # Otherwise it returns the passed argument.
 
-    val_in_bytes = None
-
-    if 'kB' in pretty_val:
-        num_part = int(''.join(d for d in pretty_val if d.isdigit()))
-        val_in_bytes = num_part * 1024
-
-    elif 'MB' in pretty_val.upper():
-        num_part = int(''.join(d for d in pretty_val if d.isdigit()))
-        val_in_bytes = num_part * 1024 * 1024
-
-    elif 'GB' in pretty_val.upper():
-        num_part = int(''.join(d for d in pretty_val if d.isdigit()))
-        val_in_bytes = num_part * 1024 * 1024 * 1024
-
-    elif 'TB' in pretty_val.upper():
-        num_part = int(''.join(d for d in pretty_val if d.isdigit()))
-        val_in_bytes = num_part * 1024 * 1024 * 1024 * 1024
-
-    elif 'B' in pretty_val.upper():
-        num_part = int(''.join(d for d in pretty_val if d.isdigit()))
-        val_in_bytes = num_part
-
-    else:
+    # It's sometimes possible to have an empty values
+    if not pretty_val:
         return pretty_val
 
-    return val_in_bytes
+    # If the first char is not a digit, it does not make sense
+    # to parse further, so just return a passed value
+    if not pretty_val[0].isdigit():
+        return pretty_val
+
+    # If the last char is not an alphabetical symbol, it means that
+    # it does not contain any suffixes, so no sense to parse further
+    if not pretty_val[-1].isalpha():
+        try:
+            pretty_val = int(pretty_val)
+
+        except ValueError:
+            pretty_val = float(pretty_val)
+
+        return pretty_val
+
+    # Extract digits
+    num_part = []
+    for c in pretty_val:
+        # When we reach the first non-digit element,
+        # e.g. in 1024kB, stop iterating
+        if not c.isdigit():
+            break
+        else:
+            num_part.append(c)
+
+    num_part = int(''.join(num_part))
+
+    val_in_bytes = None
+
+    if len(pretty_val) >= 2:
+        if 'kB' in pretty_val[-2:]:
+            val_in_bytes = num_part * 1024
+
+        elif 'MB' in pretty_val[-2:]:
+            val_in_bytes = num_part * 1024 * 1024
+
+        elif 'GB' in pretty_val[-2:]:
+            val_in_bytes = num_part * 1024 * 1024 * 1024
+
+        elif 'TB' in pretty_val[-2:]:
+            val_in_bytes = num_part * 1024 * 1024 * 1024 * 1024
+
+    # For cases like "1B"
+    if not val_in_bytes and 'B' in pretty_val[-1]:
+        val_in_bytes = num_part
+
+    if val_in_bytes is not None:
+        return val_in_bytes
+    else:
+        return pretty_val
 
 
 def param_set(cursor, module, name, value, context):
@@ -289,7 +331,7 @@ def main():
         name=dict(type='str', required=True),
         db=dict(type='str', aliases=['login_db']),
         value=dict(type='str'),
-        reset=dict(type='bool'),
+        reset=dict(type='bool', default=False),
         session_role=dict(type='str'),
         trust_input=dict(type='bool', default=True),
     )
@@ -308,11 +350,14 @@ def main():
         # Check input for potentially dangerous elements:
         check_input(module, name, value, session_role)
 
-    # Allow to pass values like 1mb instead of 1MB, etc:
     if value:
-        for unit in POSSIBLE_SIZE_UNITS:
-            if value[:-2].isdigit() and unit in value[-2:]:
-                value = value.upper()
+        # Convert a value like 1mb (Postgres does not support) to 1MB, etc:
+        if len(value) > 2 and value[:-2].isdigit() and value[-2:] in LOWERCASE_SIZE_UNITS:
+            value = value.upper()
+
+        # Convert a value like 1b (Postgres does not support) to 1B:
+        elif len(value) > 1 and ('b' in value[-1] and value[:-1].isdigit()):
+            value = value.upper()
 
     if value is not None and reset:
         module.fail_json(msg="%s: value and reset params are mutually exclusive" % name)