+++ ansible_collections/community/postgresql/plugins/modules/postgresql_slot.py
@@ -77,6 +77,7 @@ options:
 notes:
 - Physical replication slots were introduced to PostgreSQL with version 9.4,
   while logical replication slots were added beginning with version 10.0.
+- Supports C(check_mode).
 
 seealso:
 - name: PostgreSQL pg_replication_slots view reference
@@ -132,7 +133,7 @@ EXAMPLES = r'''
 
 RETURN = r'''
 name:
-  description: Name of the slot
+  description: Name of the slot.
   returned: always
   type: str
   sample: "physical_one"
@@ -190,8 +191,8 @@ class PgSlot(object):
             return None
 
         if kind == 'physical':
-            # Check server version (needs for immedately_reserverd needs 9.6+):
-            if self.cursor.connection.server_version < 96000:
+            # Check server version (immediately_reserved needs 9.6+):
+            if self.cursor.connection.server_version < 90600:
                 query = "SELECT pg_create_physical_replication_slot(%(name)s)"
 
             else: